requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Sastrawi>=1.0.1
nltk>=3.8.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
lxml>=4.9.0
joblib>=1.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
